feat: deprecate `PseudoBox` and replace all `PseudoBox` components with `Box`
