+++ packages/react-styled-ui/src/Modal/ModalBody.js
@@ -1,5 +1,5 @@
 import React, { forwardRef } from 'react';
-import PseudoBox from '../PseudoBox';
+import Box from '../Box';
 import {
   useModalBodyStyle,
 } from './styles';
@@ -8,7 +8,7 @@ const ModalBody = forwardRef((props, ref) => {
   const styleProps = useModalBodyStyle();
 
   return (
-    <PseudoBox
+    <Box
       ref={ref}
       {...styleProps}
       {...props}
