+++ packages/react-styled-ui/src/Scrollbar/renderDefaultElements.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import Box from '../Box';
-import PseudoBox from '../PseudoBox';
 
 const renderViewDefault = (props) => {
   return <Box {...props} />;
@@ -15,11 +14,11 @@ const renderTrackVerticalDefault = (props) => {
 };
 
 const renderThumbHorizontalDefault = (props) => {
-  return <PseudoBox {...props} />;
+  return <Box {...props} />;
 };
 
 const renderThumbVerticalDefault = (props) => {
-  return <PseudoBox {...props} />;
+  return <Box {...props} />;
 };
 
 export {
