+++ packages/styled-ui-docs/pages/alert.mdx
@@ -94,7 +94,7 @@ Setting the `icon` prop to `false` will remove the icon altogether.
 
 ```jsx noInline
 const Paragraph = (props) => (
-  <PseudoBox mb="1x" _lastChild={{ mb: 0 }} {...props} />
+  <Box mb="1x" _lastChild={{ mb: 0 }} {...props} />
 );
 
 function Example() {
